Add linters fix, logic separation
