+++ src/logic.js
@@ -5,9 +5,10 @@ const weatherDataFilter = (data, location) => {
   const weatherIcon = data.weather[0].icon;
   const realFeel = data.main.feels_like;
   const weatherBackground = data.weather[0].main;
-  const fCondition = Math.floor(condition * (9 / 5) + 32);
+  const fCondition = (condition * (9 / 5) + 32).toFixed(2);
+  const fRealFeel = (realFeel * (9 / 5) + 32).toFixed(2);
   return {
-    cityName, condition, weather, weatherIcon, realFeel, weatherBackground, fCondition,
+    cityName, condition, weather, weatherIcon, realFeel, weatherBackground, fCondition, fRealFeel,
   };
 };
